+++ test/daisy_ui_components/label_test.exs
@@ -1,87 +1,111 @@
 defmodule DaisyUIComponents.LabelTest do
-  use ExUnit.Case
+  use DaisyUIComponents.ComponentCase
 
   import Phoenix.Component
-  import Phoenix.LiveViewTest
   import DaisyUIComponents.Label
 
   test "label" do
     assigns = %{}
 
-    label =
-      rendered_to_string(~H"""
-      <.label>label slot</.label>
-      """)
-
-    assert label =~ ~s(<label class="label">)
-    assert label =~ ~s(label slot)
+    ~H"""
+    <.label class="w-xs">label slot</.label>
+    """
+    |> parse_component()
+    |> assert_component("label")
+    |> assert_class("label w-xs")
+    |> assert_text("label slot")
   end
 
   test "label input" do
     assigns = %{}
 
-    assert label =
-             rendered_to_string(~H"""
-             <.label type="input" text="input label">
-               <input type="text" placeholder="URL" />
-             </.label>
-             """)
-
-    assert label =~ ~s(<label class="input">)
-    assert label =~ ~s(<span class="label">)
-    assert label =~ ~s(input label)
-    assert label =~ ~s(<input type="text" placeholder="URL">)
+    ~H"""
+    <.label type="input" text="input label" class="w-xs">
+      <input type="text" placeholder="URL" />
+    </.label>
+    """
+    |> parse_component()
+    |> assert_component("label")
+    |> assert_class("input w-xs")
+    |> select_element("span", fn element ->
+      element
+      |> assert_component("span")
+      |> assert_text("input label")
+    end)
+    |> select_element("input", fn element ->
+      element
+      |> assert_component("input")
+    end)
   end
 
   test "label select" do
     assigns = %{}
 
-    assert label =
-             rendered_to_string(~H"""
-             <.label type="select" text="input label">
-               <select />
-             </.label>
-             """)
-
-    assert label =~ ~s(<label class="select">)
-    assert label =~ ~s(<span class="label">)
-    assert label =~ ~s(input label)
-    assert label =~ ~s(<select>)
+    ~H"""
+    <.label type="select" text="input label" class="w-xs">
+      <select />
+    </.label>
+    """
+    |> parse_component()
+    |> assert_component("label")
+    |> assert_class("select w-xs")
+    |> select_element("span", fn element ->
+      element
+      |> assert_component("span")
+      |> assert_text("input label")
+    end)
+    |> select_element("select", fn element ->
+      element
+      |> assert_component("select")
+    end)
   end
 
   test "label floating" do
     assigns = %{}
 
-    assert label =
-             rendered_to_string(~H"""
-             <.label type="floating" text="Your email">
-               <input type="text" placeholder="mail@site.com" class="input input-md" />
-             </.label>
-             """)
-
-    assert label =~ ~s(<label class="floating-label">)
-    assert label =~ ~s(<span class="">)
-    assert label =~ ~s(Your email)
-    assert label =~ ~s(<input type="text" placeholder="mail@site.com" class="input input-md">)
+    ~H"""
+    <.label type="floating" text="Your email">
+      <input type="text" placeholder="mail@site.com" class="input input-md" />
+    </.label>
+    """
+    |> parse_component()
+    |> assert_component("label")
+    |> assert_class("floating-label")
+    |> select_element("span", fn element ->
+      element
+      |> assert_component("span")
+      |> assert_text("Your email")
+    end)
+    |> select_element("input", fn element ->
+      element
+      |> assert_component("input")
+    end)
   end
 
   test "label positions" do
     assigns = %{}
 
-    assert label =
-             rendered_to_string(~H"""
-             <.label type="input" text=".com" position="end">
-               <input type="text" placeholder="domain name" />
-             </.label>
-             """)
-
-    {:ok, document} = Floki.parse_document(label)
-    [{"label", _attr, child}] = Floki.find(document, "label")
-    # assert order of elements
-    assert [{"input", _, _}, {"span", _, _}] = child
-
-    assert label =~ ~s(<label class="input">)
-    assert label =~ ~s(<span class="label">)
-    assert label =~ ~s(<input type="text" placeholder="domain name">)
+    ~H"""
+    <.label type="input" text=".com" position="end">
+      <input type="text" placeholder="domain name" />
+    </.label>
+    """
+    |> parse_component()
+    |> assert_component("label")
+    |> assert_class("input")
+    |> select_element("span", fn element ->
+      element
+      |> assert_component("span")
+      |> assert_text(".com")
+    end)
+    |> select_element("input", fn element ->
+      element
+      |> assert_component("input")
+    end)
+    |> select_element("label", fn element ->
+      [{"label", _attr, child}] = element
+      # assert order of elements
+      assert [{"input", _, _}, {"span", _, _}] = child
+    end)
   end
 end