+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 - Label storybook
 - Fieldset component
 - Fieldset storybook
+- Progress component
+- Progress storybook
 
 ### Changed